Symmetric naming of Esri JSON and GeoJSON - InputFormat/RecordReader (#81)
